pyproject.toml
[project]
name = "hr-resume-screener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyterlab>=4.3.5",
    "pymupdf>=1.25.2",
    "python-dotenv>=1.0.1",
    "voyageai>=0.3.2",
]
